Feature functional
This feature relies on the global variable _targetFeatureList which is a
2d array storing the point indexes of the left and right points
corresponding to key facial features. It uses a vector based approach
comparing vector lengths between the bridge of the nose and the feature.

This process could be improved by also calcluating the angle of the
vectors in relation to the orientation of the face. Additionally, the
colour augmentations and displayed nodes could include some of those
points surrounding the target features. By iterating from their indexes
we can display and colour adjacent nodes as well as our targets.